Move common DBus proxy code to a single base class
Also consolidate into a single file.
