+++ blueman/plugins/services/Transfer.py
@@ -4,7 +4,7 @@ import logging
 
 from blueman.Constants import *
 from blueman.plugins.ServicePlugin import ServicePlugin
-from blueman.main.AppletService import AppletService
+from blueman.main.DBusProxies import AppletService
 from blueman.main.Config import Config
 
 import gi
