+++ po/POTFILES.in
@@ -11,16 +11,15 @@
 [type: gettext/glade]data/ui/net-usage.ui
 [type: gettext/glade]data/ui/note.ui
 blueman/main/Adapter.py
+blueman/main/DBusProxies.py
 blueman/main/Manager.py
 blueman/main/Sendto.py
 blueman/main/Services.py
 blueman/main/NetConf.py
 blueman/main/PulseAudioUtils.py
-blueman/main/Mechanism.py
 blueman/main/PPPConnection.py
 blueman/main/Config.py
 blueman/main/__init__.py
-blueman/main/AppletService.py
 blueman/main/DhcpClient.py
 blueman/main/SpeedCalc.py
 blueman/main/applet/__init__.py
